+++ src/UserGameMode.java
@@ -1,5 +1,3 @@
-package initializer;
-
 import java.util.HashMap;
 import java.util.Map;
