+++ src/session/Winner.java
@@ -1,0 +1,8 @@
+package session;
+
+public enum Winner {
+    DRAW,
+    FIRST_PLAYER,
+    SECOND_PLAYER,
+    UNKNOWN
+}
